+++ api/partimark_app/schemas/system_config.py
@@ -10,7 +10,7 @@ from pydantic import BaseModel, ConfigDict, Field
 class SystemConfigBase(BaseModel):
     """Shared properties across system configuration schemas."""
     coordinator_user_id: str = Field(..., max_length=50)
-    max_weekly_score: int = Field(3, ge=0)
+    max_weekly_score: float = Field(3, ge=0)
     total_participation_points: int = Field(0, ge=0)
     is_configured: bool = False
     week6_lock_enabled: bool = False
@@ -37,7 +37,7 @@ class SystemConfigUpdate(BaseModel):
     All fields are optional to support partial updates.
     """
     coordinator_user_id: Optional[str] = Field(None, max_length=50)
-    max_weekly_score: Optional[int] = Field(None, ge=0)
+    max_weekly_score: Optional[float] = Field(None, ge=0)
     total_participation_points: Optional[int] = Field(None, ge=0)
     is_configured: Optional[bool] = None
     week6_lock_enabled: Optional[bool] = None
